fix invalid optimization for nop nodes (#137)
Graph was invalid in some cases when trying to optimize nop nodes (nodes
that just does nothing like empty targets).

If the builder decides not to create a node (because the node does
nothing), it now store dependencies appart to reuse them. Previously,
the builder was reporting nop nodes, leading to stronger dependencies as
required (and sometimes wrong dependencies).