feat: cacheability update (ephemeral & cache) (#209)
# ephemeral
add implement for `ephemeral` target. This replaces the previous concept
of `managed` target.

Ephemeral attribute can be applied on target. The default value is
`false` unless the last task in a target is ephemeral.
Ephemeral target are always re-evaluated on `retry`.

# cache
Before, cache was determined as the intersection of cacheability of all
tasks of a target.
Now, the last target wins.