feat: extension owned project identifier (#303)
Terrabuild generates the project id by itself. It's based on the
relative path of the project. This has been a pain with pnpm and
workspaces.

This PR introduces project based identifier, allowing extension to
managed own references. If the extension is not aware or does not
require this, Terrabuild works as before and generate appropriate
identifier.

Note this PR also deprecates `references` attribute on project - as this
was file system based. If required to add dependencies, use
`depends_on`.